refactor: Emitter renamed to streamer and separated from path module
